requirements.txt
anyio==3.5.0
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.10.0
boto3==1.21.21
botocore==1.24.21
cached-property==1.5.2
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.4
cryptography==36.0.2
datapackage==1.15.2
deprecation==2.1.0
et-xmlfile==1.1.0
gotrue==0.5.0
greenlet==1.1.2
h11==0.12.0
httpcore==0.14.7
httpx==0.21.3
idna==3.3
ijson==3.1.4
importlib-resources==5.4.0
iniconfig==1.1.1
isodate==0.6.1
jmespath==1.0.0
jsonlines==3.0.0
jsonpointer==2.2
jsonschema==4.4.0
linear-tsv==1.1.0
numpy==1.22.3
openpyxl==3.0.9
outcome==1.1.0
packaging==21.3
pandas==1.4.1
pluggy==1.0.0
postgrest-py==0.9.2
py==1.11.0
pycparser==2.21
pydantic==1.9.0
pyOpenSSL==22.0.0
pyparsing==3.0.7
pyrsistent==0.18.1
PySocks==1.7.1
pytest==7.1.1
pytest-xvfb==2.0.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyVirtualDisplay==3.0
realtime==0.0.4
requests==2.27.1
rfc3986==1.5.0
s3transfer==0.5.2
selenium==4.1.3
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
SQLAlchemy==1.4.32
supabase==0.5.3
tableschema==1.20.2
tabulator==1.53.5
tomli==2.0.1
trio==0.20.0
trio-websocket==0.9.2
typing-extensions==4.1.1
unicodecsv==0.14.1
urllib3==1.26.9
websockets==9.1
wsproto==1.1.0
xlrd==2.0.1
zipp==3.7.0
